requirements.txt
influxdb~=5.3.1
bme280~=0.6
psutil~=5.9.5
fritzconnection~=1.12.0
matrix_nio~=0.20.2
numexpr~=2.8.4